% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalise.R
\name{capitalise}
\alias{capitalise}
\title{Capitalise function}
\usage{
capitalise(x)
}
\arguments{
\item{x}{column to be manipulated}
}
\value{
None
}
\description{
Function to capitalise species names from \code{\link{GlobalArchive}}
}
\examples{
capitalise("tEst dATa")

}
