% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{change.synonyms}
\alias{change.synonyms}
\title{change.synonyms function}
\usage{
change.synonyms(dat, return.changes = FALSE)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Change all synonyms \code{\link{GlobalArchive}}
}
\examples{
change.synonyms()

}
