% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.check_api.R
\name{check.api}
\alias{check.api}
\title{check.api function}
\usage{
check.api(API_USER_TOKEN)
}
\arguments{
\item{API_USER_TOKEN}{needs user token to connect to \code{\link{GlobalArchive}}}
}
\value{
None
}
\description{
Function that checks connection to \code{\link{GlobalArchive}}
}
\examples{
check.api()

}
