% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean names function}
\usage{
clean_names(dat)
}
\arguments{
\item{dat}{data to be renamed}
}
\value{
None
}
\description{
Creates lower case names with no special characters in \code{\link{GlobalArchive}}
}
\examples{
clean_names()

}
