% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.length3dpoints.R
\name{create.length3dpoints}
\alias{create.length3dpoints}
\title{create.length3dpoints function}
\usage{
create.length3dpoints(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
The create.length3dpoints function finds all files ending in "Lengths.txt","3DPoints.txt" and "Count.csv" and combines them into one dataframe \code{\link{GlobalArchive}}
}
\examples{
create.length3dpoints()

}
