% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.maxn.R
\name{create.maxn}
\alias{create.maxn}
\title{create.maxn function}
\usage{
create.maxn(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
The create.maxn function finds all files ending in "Points.txt" and "Count.csv" and combines them into a maxn file \code{\link{GlobalArchive}}
}
\examples{

}
