% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.files.R
\name{expand.files}
\alias{expand.files}
\title{expand.files function}
\usage{
\method{expand}{files}(files)
}
\arguments{
\item{dat}{}
}
\value{
None
}
\description{
Function that expands list of files as a data.frame from \code{\link{GlobalArchive}}
}
\examples{
expand.files()

}
