% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.capitalise.R
\name{ga.capitalise}
\alias{ga.capitalise}
\title{Capitalise function}
\usage{
ga.capitalise(x)
}
\arguments{
\item{x}{column to be manipulated}
}
\value{
None
}
\description{
Function to capitalise species names from \code{\link{GlobalArchive}}
}
\examples{
ga.capitalise("tEst dATa")

}
