% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.synonyms.R
\name{ga.change.synonyms}
\alias{ga.change.synonyms}
\title{change.synonyms function}
\usage{
ga.change.synonyms(dat, return.changes = FALSE, save.report = FALSE)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Change all synonyms \code{\link{GlobalArchive}}
}
\examples{

}
