% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.clean.names.R
\name{ga.clean.names}
\alias{ga.clean.names}
\title{Clean names function}
\usage{
ga.clean.names(dat)
}
\arguments{
\item{dat}{data to be renamed}
}
\value{
None
}
\description{
Creates lower case names with no special characters in \code{\link{GlobalArchive}}
}
\examples{

}
