% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.create.em.maxn.R
\name{ga.create.em.maxn}
\alias{ga.create.em.maxn}
\title{ga.create.em.maxn function}
\usage{
ga.create.em.maxn(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
The ga.create.em.maxn function finds all files ending in "Points.txt" and combines them into a maxn file, keeps filename, and periodtime DOES NOT include stage
}
\examples{

}
