% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.create.length3dpoints.R
\name{ga.create.length3dpoints}
\alias{ga.create.length3dpoints}
\title{ga.create.length3dpoints function}
\usage{
ga.create.length3dpoints(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
The ga.create.length3dpoints function finds all files ending in "Lengths.txt","3DPoints.txt" and "Count.csv" and combines them into one dataframe \code{\link{GlobalArchive}}
}
\examples{

}
