% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.create.maxn.R
\name{ga.create.maxn}
\alias{ga.create.maxn}
\title{ga.create.maxn function}
\usage{
ga.create.maxn(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
The ga.create.maxn function finds all files ending in "Points.txt" and "Count.csv" and combines them into a maxn file, keeps filename, and periodtime DOES NOT include stage
}
\examples{

}
