% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.download.campaign_file.R
\name{ga.download.campaign_file}
\alias{ga.download.campaign_file}
\title{ga.download.campaign_file function}
\usage{
ga.download.campaign_file(api_token, id, save_file_path)
}
\arguments{
\item{api_token}{api_token}

\item{id}{id}

\item{save_file_path}{save_file_path}
}
\value{
prints output
}
\description{
Function that dowloads particular campaign from  \code{\link{GlobalArchive}}
}
\examples{

}
