% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.download.campaign_files.R
\name{ga.download.campaign_files}
\alias{ga.download.campaign_files}
\title{ga.download.campaign_files function}
\usage{
ga.download.campaign_files(api_token, campaign_files, save_path,
  match = NULL, exclude = NULL)
}
\arguments{
\item{api_token}{api_token}

\item{campaign_files}{campaign_files}

\item{save_path}{api_token}
}
\value{
prints output
}
\description{
Function that downloads campaign_files from  \code{\link{GlobalArchive}}
}
\examples{

}
