% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.download.campaign_info.R
\name{ga.download.campaign_info}
\alias{ga.download.campaign_info}
\title{ga.download.campaign_info function}
\usage{
ga.download.campaign_info(api_token, campaign_info, save_path)
}
\arguments{
\item{api_token}{api_token}

\item{campaign_info}{campaign_info}

\item{save_path}{save_path}
}
\value{
prints output
}
\description{
Function that downloads particular campaign info from  \code{\link{GlobalArchive}}
}
\examples{

}
