% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.download.campaign_record.R
\name{ga.download.campaign_record}
\alias{ga.download.campaign_record}
\title{ga.download.campaign_record function}
\usage{
ga.download.campaign_record(api_token, campaign_details, save_path)
}
\arguments{
\item{api_token}{api_token}

\item{campaign_details}{campaign_details}

\item{save_file_path}{save_file_path}
}
\value{
prints output
}
\description{
Function that dowloads particular campaign record from  \code{\link{GlobalArchive}}
}
\examples{
ga.download.campaign_record()

}
