% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.download.file.R
\name{ga.download.file}
\alias{ga.download.file}
\title{ga.download.file function}
\usage{
ga.download.file(api_token, api_endpoint, path, query = NULL)
}
\arguments{
\item{api_token}{api_token}

\item{api_endpoint}{api_endpoint}

\item{path}{to where results will be saved}
}
\value{
A dataframe containing the json response from the API
}
\description{
Function that performs paginated multiple HTTP GET requests on api endpoint and processes each object with optional processing function from  \code{\link{GlobalArchive}}
}
\examples{
process_campaign_object()

}
