% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.get.campaign.list.R
\name{ga.get.campaign.list}
\alias{ga.get.campaign.list}
\title{ga.get.campaign.list function}
\usage{
ga.get.campaign.list(api_token, f, q = "")
}
\arguments{
\item{api_token}{api_token}
}
\value{
prints list of campaigns
}
\description{
Function that gets campaign list from  \code{\link{GlobalArchive}}
}
\examples{

}
