% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.get.campaign.R
\name{ga.get.campaign}
\alias{ga.get.campaign}
\title{ga.get.campaign function}
\usage{
ga.get.campaign(api_token, campaign_id)
}
\arguments{
\item{api_token}{api_token}
}
\value{
null
}
\description{
Function that gets campaign from  \code{\link{GlobalArchive}}
}
\examples{
ga.get.campaign()

}
