% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.get.object_list.R
\name{ga.get.object_list}
\alias{ga.get.object_list}
\title{ga.get.object_list function}
\usage{
ga.get.object_list(api_endpoint, api_token, f = NULL, q = "")
}
\arguments{
\item{api_endpoint}{api_endpoint}

\item{api_token}{api_token}

\item{f}{f optional, default=NULL, function pointer to process each object. If NULL, do nothing}

\item{q}{optional, default=NULL: json query string for GA API}
}
\value{
A dataframe containing the json response from the API
}
\description{
Function that performs paginated multiple HTTP GET requests on api endpoint and processes each object with optional processing function from  \code{\link{GlobalArchive}}
}
\examples{

}
