% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.length.sp.to.spp.R
\name{ga.lengths.sp.to.spp}
\alias{ga.lengths.sp.to.spp}
\title{ga.lengths.sp.to.spp function}
\usage{
ga.lengths.sp.to.spp(dat, sp.list, return.changes = FALSE)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Change all sp1 sp2 sp3 to spp and only keep the one with the most length measurements \code{\link{GlobalArchive}}
}
\examples{

}
