% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.list.files.R
\name{ga.list.files}
\alias{ga.list.files}
\title{ga.list.files function}
\usage{
ga.list.files(pattern)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Function that lists campaign objects from \code{\link{GlobalArchive}}
}
\examples{

}
