% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.print.campaign_details.R
\name{ga.print.campaign_details}
\alias{ga.print.campaign_details}
\title{ga.print.campaign_details function}
\usage{
ga.print.campaign_details(campaign_details)
}
\arguments{
\item{campaign_details}{campaign_details}
}
\value{
prints output
}
\description{
Function that prints particular campaign info from  \code{\link{GlobalArchive}}
}
\examples{

}
