% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.print.R
\name{ga.print}
\alias{ga.print}
\title{ga.print function}
\usage{
ga.print(format, ...)
}
\arguments{
\item{format}{format}
}
\value{
prints output
}
\description{
Function that prints objects from  \code{\link{GlobalArchive}}
}
\examples{

}
