% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.query.campaign.R
\name{ga.query.campaign}
\alias{ga.query.campaign}
\title{ga.query.campaign function}
\usage{
ga.query.campaign(campaign)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by campaign name \code{\link{GlobalArchive}}
}
\examples{

}
