% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.query.pattern.R
\name{ga.query.pattern}
\alias{ga.query.pattern}
\title{ga.query.pattern function}
\usage{
ga.query.pattern(pattern)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by pattern name \code{\link{GlobalArchive}}
}
\examples{

}
