% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.query.project.R
\name{ga.query.project}
\alias{ga.query.project}
\title{ga.query.project function}
\usage{
ga.query.project(project)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by project name \code{\link{GlobalArchive}}
}
\examples{

}
