% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.query.workgroup.R
\name{ga.query.workgroup}
\alias{ga.query.workgroup}
\title{ga.query.workgroup function}
\usage{
ga.query.workgroup(workgroup)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by workgroup name \code{\link{GlobalArchive}}
}
\examples{

}
