% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.read.files_csv.R
\name{ga.read.files_csv}
\alias{ga.read.files_csv}
\title{Read csv function}
\usage{
ga.read.files_csv(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in csv files and creates a column for filepath to get CampaignID from \code{\link{GlobalArchive}}
}
\examples{

}
