% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.read.files_em.csv.R
\name{ga.read.files_em.csv}
\alias{ga.read.files_em.csv}
\title{Read EventMeasure csv function}
\usage{
ga.read.files_em.csv(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in csv files and creates a column for the file name to get CampaignID from EventMeasure database tables
}
\examples{

}
