% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.read.files_em.txt.R
\name{ga.read.files_em.txt}
\alias{ga.read.files_em.txt}
\title{Read EventMeasure txt function}
\usage{
ga.read.files_em.txt(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in txt files and creates a column for filename to get CampaignID from EventMeasure database tables
}
\examples{

}
