% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.read.files_txt.R
\name{ga.read.files_txt}
\alias{ga.read.files_txt}
\title{Read txt function}
\usage{
ga.read.files_txt(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in txt files and creates a column for filepath to get CampaignID from \code{\link{GlobalArchive}}
}
\examples{

}
