% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.taxa.replaced.by.spp.R
\name{ga.taxa.replaced.by.spp}
\alias{ga.taxa.replaced.by.spp}
\title{taxa.replaced.by.spp function}
\usage{
ga.taxa.replaced.by.spp(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
List all taxa replaced by all sp1 sp2 sp3 to spp \code{\link{GlobalArchive}}
}
\examples{

}
