% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy.birthday.todd.R
\name{happy.birthday.todd}
\alias{happy.birthday.todd}
\title{happy.birthday.todd function}
\usage{
happy.birthday.todd(dat)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Happy Bday Todd!
}
\examples{

}
