% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.sp.to.spp.R
\name{length.sp.to.spp}
\alias{length.sp.to.spp}
\title{length.sp.to.spp function}
\usage{
\method{length}{sp.to.spp}(dat, sp.list, return.changes = FALSE)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Change all sp1 sp2 sp3 to spp \code{\link{GlobalArchive}}
}
\examples{
length.sp.to.spp()

}
