% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.files.GA.R
\name{list.files.GA}
\alias{list.files.GA}
\title{list.files.GA function}
\usage{
list.files.GA(pattern)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Function that lists campaign objects from \code{\link{GlobalArchive}}
}
\examples{
list.files.GA()

}
