% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxn.sp.to.spp.R
\name{maxn.sp.to.spp}
\alias{maxn.sp.to.spp}
\title{Sp to spp function}
\usage{
maxn.sp.to.spp(dat, sp.list, return.changes = FALSE)
}
\arguments{
\item{dat}{file to be manipulated}

\item{sp.list}{list of species to compare with}
}
\value{
None
}
\description{
Function that corrects sp to spp based on a species list from \code{\link{GlobalArchive}}
}
\examples{
maxn.sp.to.spp()

}
