% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.process_campaign_object.R
\name{process_campaign_object}
\alias{process_campaign_object}
\title{process_campaign_object function}
\usage{
process_campaign_object(object)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Function that returns campaign objects from \code{\link{GlobalArchive}}
}
\examples{
process_campaign_object()

}
