% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.campaign.R
\name{query.campaign}
\alias{query.campaign}
\title{query.campaign function}
\usage{
query.campaign(campaign)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by campaign name \code{\link{GlobalArchive}}
}
\examples{

}
