% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.pattern.R
\name{query.pattern}
\alias{query.pattern}
\title{query.pattern function}
\usage{
query.pattern(pattern)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by pattern name \code{\link{GlobalArchive}}
}
\examples{
query.pattern()

}
