% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.project.R
\name{query.project}
\alias{query.project}
\title{query.project function}
\usage{
query.project(project)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by project name \code{\link{GlobalArchive}}
}
\examples{
query.project()

}
