% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.workgroup.R
\name{query.workgroup}
\alias{query.workgroup}
\title{query.workgroup function}
\usage{
query.workgroup(workgroup)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Query by workgroup name \code{\link{GlobalArchive}}
}
\examples{
query.workgroup()

}
