% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_csv.R
\name{read_files_csv}
\alias{read_files_csv}
\title{Read csv function}
\usage{
read_files_csv(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in csv files and creates a column for filepath to get CampaignID from \code{\link{GlobalArchive}}
}
\examples{
read_files_csv()

}
