% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_txt.R
\name{read_files_txt}
\alias{read_files_txt}
\title{Read txt function}
\usage{
read_files_txt(flnm)
}
\arguments{
\item{flnm}{file to be manipulated}
}
\value{
None
}
\description{
Function that reads in txt files and creates a column for filepath to get CampaignID from \code{\link{GlobalArchive}}
}
\examples{
read_files_txt()

}
