% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_https.R
\name{source_https}
\alias{source_https}
\title{source_https function}
\usage{
source_https(url, ...)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Function that brings in url values from github \code{\link{GlobalArchive}}
}
\examples{
list.files.GA()

}
