% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalArchive-package.R
\docType{package}
\name{GlobalArchive-package}
\alias{GlobalArchive}
\alias{GlobalArchive-package}
\title{GlobalArchive: Access GlobalArchive data}
\description{
Access GlobalArchive data via API.
}
\author{
\strong{Maintainer}: Brooke Gibbons \email{brooke.gibbons@uwa.edu.au} (\href{https://orcid.org/0000-0002-0628-7781}{ORCID})

Authors:
\itemize{
  \item Tim Langlois (\href{https://orcid.org/0000-0001-6404-4000}{ORCID})
}

Other contributors:
\itemize{
  \item ARDC [funder]
  \item NESP MAC [funder]
}

}
\keyword{internal}
