% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.files.GA.R
\name{list.files.GA}
\alias{list.files.GA}
\title{List files in GA}
\usage{
list.files.GA(pattern)
}
\arguments{
\item{dat}{data to be renamed}
}
\value{
None
}
\description{
Read in files that match the pattern in your downloads dir
}
\examples{
list.files.GA()

}
