% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxn.sp.to.spp.R
\name{maxn.sp.to.spp}
\alias{maxn.sp.to.spp}
\title{maxn.sp.to.spp function}
\usage{
maxn.sp.to.spp(dat, sp.list, return.changes = FALSE)
}
\arguments{
\item{object}{file to be manipulated}
}
\value{
None
}
\description{
Change all sp1 sp2 sp3 to spp \code{\link{GlobalArchive}}
}
\examples{

}
